Reconstructed file structure and solved most online issues
